+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   support ([#44](https://github.com/gobley/gobley/pull/44) & [#47](https://github.com/gobley/gobley/pull/47)).
 - Config option for using PascalCase for enums ([#54](https://github.com/gobley/gobley/pull/54)).
 - `CargoCheckTask` for cross-platform linting ([#55](https://github.com/gobley/gobley/pull/55)).
+- `UniFfiPlugin` can be used without a UniFFI config
+  file ([#63](https://github.com/gobley/gobley/pull/63)).
 
 ### Fixes